Fix remaining method parameters that incorrectly specify the method's parent type instead of their true buffer type

+++ ManagedCorDebug/Native/Interfaces/ICLRMetadataLocator.cs
@@ -19,7 +19,7 @@ namespace ManagedCorDebug
             [In] uint mdRva,
             [In] uint flags,
             [In] uint bufferSize,
-            out byte buffer,
+            out IntPtr buffer,
             out uint dataSize);
     }
 }
